refactor: #656 interview メディア/セッション hooks を抽出する
page を薄いオーケストレーションにし、stream・VAD・セッション生命周期を
hooks へ集約する。refs の所有は useInterviewMedia に一本化する。
